backend/requirements.txt
absl-py==2.3.1
altgraph==0.17.4
Authlib==1.6.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Session==0.8.0
idna==3.10
immutabledict==4.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
numpy==2.3.2
openpyxl==3.1.5
ortools==9.14.6206
packaging==25.0
pandas==2.3.1
pefile==2023.2.7
protobuf==6.31.1
pycparser==2.22
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
redis==6.4.0
requests==2.32.4
setuptools==80.9.0
six==1.17.0
tabulate==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
